requirements.txt
opencv-python-headless<4.3
Pillow
gradio
torch
easyocr